fixed issues with Jackson & numeric type widening

+++ app/v1/gth/render/RenderingParams.scala
@@ -1,7 +1,7 @@
 package v1.gth.render
 
 case class RenderingParams(
-                            fontSize: Byte = 12,
-                            margin: Byte = 5,
-                            lineWidth: Byte = 1
+                            fontSize: Double = 12,
+                            margin: Double = 5,
+                            lineWidth: Double = 1
                           )
